ДЗ Модуляция
module5-task1
