+++ js/data-users.js
@@ -1,0 +1,24 @@
+// Массив имён
+const USERS_NAMES = [
+  'Александр', 'Дмитрий', 'Максим', 'Сергей', 'Андрей', 'Алексей', 'Артём', 'Илья', 'Кирилл', 'Михаил', 'Никита', 'Анастасия', 'Мария', 'Анна', 'Виктория', 'Екатерина', 'Наталья', 'Марина', 'Полина', 'София', 'Дарья', 'Алиса', 'Ксения'
+];
+
+// Массив коментов
+const USERS_COMMENTS = [
+  'Всё отлично!, В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+// Массив описания фотографий
+const USERS_DESCRIPTIONS = [
+  'Мейн-кун - Одной из самых удивительных и загадочных пород считается мейн-кун – ласковый гигант с серьезным взглядом.',
+  'Шотландская вислоухая - Во всём мире эту породу именуют «скоттиш-фолд», но нам привычней называть этих милых кошек «шотландскими вислоухими».',
+  'Британские короткошерстные - Без таких вот «плюшевых» комочков счастья не мыслят своего существования миллионы людей во всем мире.',
+  'Бенгальские кошки – Невероятно красивые животные, которые сочетают в себе эффектную внешность диких хищников и покладистый характер домашних питомцев.',
+  'Сиамы – грациозные восточные красавцы и красавицы (сумевшие сохранить свои корни практически в первозданном виде) выделяются особым изяществом, темпераментным характером и некой миниатюрностью.'
+];
+
+export {USERS_NAMES, USERS_COMMENTS, USERS_DESCRIPTIONS};
